+++ core/dbt/contracts/graph/compiled.py
@@ -6,6 +6,7 @@ from dbt.contracts.graph.parsed import (
     ParsedHookNode,
     ParsedModelNode,
     ParsedExposure,
+    ParsedMetric,
     ParsedResource,
     ParsedRPCNode,
     ParsedSqlNode,
@@ -232,8 +233,10 @@ CompileResultNode = Union[
     ParsedSourceDefinition,
 ]
 
-# anything that participates in the graph: sources, exposures, manifest nodes
+# anything that participates in the graph: sources, exposures, metrics,
+# or manifest nodes
 GraphMemberNode = Union[
     CompileResultNode,
     ParsedExposure,
+    ParsedMetric,
 ]