+++ core/dbt/node_types.py
@@ -17,6 +17,7 @@ class NodeType(StrEnum):
     Source = 'source'
     Macro = 'macro'
     Exposure = 'exposure'
+    Metric = 'metric'
 
     @classmethod
     def executable(cls) -> List['NodeType']:
@@ -49,7 +50,8 @@ class NodeType(StrEnum):
             cls.Source,
             cls.Macro,
             cls.Analysis,
-            cls.Exposure
+            cls.Exposure,
+            cls.Metric
         ]
 
     def pluralize(self) -> str: