+++ core/dbt/task/list.py
@@ -2,7 +2,8 @@ import json
 
 from dbt.contracts.graph.parsed import (
     ParsedExposure,
-    ParsedSourceDefinition
+    ParsedSourceDefinition,
+    ParsedMetric
 )
 from dbt.graph import ResourceTypeSelector
 from dbt.task.runnable import GraphRunnableTask, ManifestTask
@@ -20,6 +21,7 @@ class ListTask(GraphRunnableTask):
         NodeType.Test,
         NodeType.Source,
         NodeType.Exposure,
+        NodeType.Metric,
     ))
     ALL_RESOURCE_VALUES = DEFAULT_RESOURCE_VALUES | frozenset((
         NodeType.Analysis,
@@ -74,6 +76,8 @@ class ListTask(GraphRunnableTask):
                 yield self.manifest.sources[node]
             elif node in self.manifest.exposures:
                 yield self.manifest.exposures[node]
+            elif node in self.manifest.metrics:
+                yield self.manifest.metrics[node]
             else:
                 raise RuntimeException(
                     f'Got an unexpected result from node selection: "{node}"'
@@ -94,6 +98,11 @@ class ListTask(GraphRunnableTask):
                 # exposures are searched for by pkg.exposure_name
                 exposure_selector = '.'.join([node.package_name, node.name])
                 yield f'exposure:{exposure_selector}'
+            elif node.resource_type == NodeType.Metric:
+                assert isinstance(node, ParsedMetric)
+                # metrics are searched for by pkg.metric_name
+                metric_selector = '.'.join([node.package_name, node.name])
+                yield f'metric:{metric_selector}'
             else:
                 # everything else is from `fqn`
                 yield '.'.join(node.fqn)