+++ test/unit/test_compiler.py
@@ -124,6 +124,7 @@ class CompilerTest(unittest.TestCase):
             disabled=[],
             files={},
             exposures={},
+            metrics={},
             selectors={},
         )
 
@@ -186,6 +187,7 @@ class CompilerTest(unittest.TestCase):
             disabled=[],
             files={},
             exposures={},
+            metrics={},
             selectors={},
         )
 
@@ -260,6 +262,7 @@ class CompilerTest(unittest.TestCase):
             disabled=[],
             files={},
             exposures={},
+            metrics={},
             selectors={},
         )
 
@@ -360,6 +363,7 @@ class CompilerTest(unittest.TestCase):
             disabled=[],
             files={},
             exposures={},
+            metrics={},
             selectors={},
         )
 
@@ -451,6 +455,7 @@ class CompilerTest(unittest.TestCase):
             disabled=[],
             files={},
             exposures={},
+            metrics={},
             selectors={},
         )
 
@@ -537,6 +542,7 @@ class CompilerTest(unittest.TestCase):
             disabled=[],
             files={},
             exposures={},
+            metrics={},
             selectors={},
         )